Added: check for invalid id

+++ src/services/notesService.ts
@@ -1,5 +1,6 @@
 import { Notes } from "../models/Notes";
 import { MESSAGES } from "../utils/messages";
+import { isValidObjectId } from "../utils/helper";
 
 class notesService {
   getnotes = async () => {
@@ -13,6 +14,9 @@ class notesService {
   }
 
   async updateNote(id: string, data: any) {
+    if (!isValidObjectId(id)) {
+      throw new Error(MESSAGES.INVALID_MONGO_ID);
+    }
     const note = await Notes.findByIdAndUpdate({ _id: id }, data, {
       new: true,
     });
@@ -23,6 +27,9 @@ class notesService {
   }
 
   async deleteNote(id: string) {
+    if (!isValidObjectId(id)) {
+      throw new Error(MESSAGES.INVALID_MONGO_ID);
+    }
     const note = await Notes.findByIdAndDelete(id);
     if (!note) {
       throw new Error(MESSAGES.NOTE_NOT_AVAILABLE);
@@ -30,6 +37,9 @@ class notesService {
   }
 
   async getNote(id: string) {
+    if (!isValidObjectId(id)) {
+      throw new Error(MESSAGES.INVALID_MONGO_ID);
+    }
     const note = await Notes.findById({ _id: id });
     if (!note) {
       throw new Error(MESSAGES.NOTE_NOT_AVAILABLE);
